server/python/requirements.txt
openai
langchain
fastapi
uvicorn
python-dotenv
dropbox